Misc bugfixes, including q3bsp-rtlights and bloom. Matrix use clarifications. Working towards skeletal glsl code.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@3890 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/bspfile.h
@@ -180,7 +180,7 @@ typedef struct
 typedef struct
 {
 	int			planenum;
-	short		children[2];	// negative numbers are contents
+	int			children[2];	// negative numbers are contents
 } mclipnode_t;
 
 typedef struct texinfo_s
